future_work/qfclient/qfclient-main/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qfclient"
version = "0.1.0"
description = "Quantitative finance data client with multi-provider support"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ethan Reinhart" },
]
keywords = [
    "finance",
    "stocks",
    "crypto",
    "market-data",
    "trading",
    "quantitative",
    "ohlcv",
    "api",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.24",
    "python-dotenv>=1.0",
    "yfinance>=0.2",
    "pandas>=2.0",
    "lxml>=4.9",  # For SEC Form 4 XML parsing
]

[project.optional-dependencies]

# Development
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
qfclient = "qfclient.cli:main"

[project.urls]
Homepage = "https://github.com/oregonquantgroup/qfclient"
Documentation = "https://github.com/oregonquantgroup/qfclient#readme"
Repository = "https://github.com/oregonquantgroup/qfclient.git"
Issues = "https://github.com/oregonquantgroup/qfclient/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/qfclient"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.git",
    "/.pytest_cache",
    "/.venv",
    "/tests",
    "/dist",
    "/.env",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "strict"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require network access)",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B"]
src = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = "src"
